🚧  notion prep before normalized through cms-api
Still a lot to do here. We are getting the basic values from:
- api/notion/[...catchAll]

In `system` we did _everything_ in `catchAll` and I think that is good
but also
overkill as it makes small changes harder  seemingly and at the cost of
being
DRY it kind of makes it look like magic a bit.

Here we will be putting in separate:
- ./pages/[routeType]/index
- ./pages/[routeType]/[...catchAll]

Or Slug I guess, could reduce the need to flip-flop and determine if
- blog
- event

As those are the only ones that use date timestamp in URL

Sigh this is a lot of stuff and I should be writing this down to ensure
I am
focusing on the right things with the multiple port here with TS too.