🚧  POC: getBlogs, getBlog(catchAll) from routeTypes
TODO:
- Remove zzz-old, after moving `preview` and `clear` over
- Port and Put Cache Back for Notion

This has been ported over from previous code from the `system` repo.
Will continue to whittle it down a bit and more TypeScript.

The reason it is `catchAll` is this was based off of `next` routes to

limit serverless function creation with `[...catchAll]` instead of

individual requests to `api/notion/...` for various `routeTypes`

and `slug` listings.

However... this may not be very straight forward still of course.