👷  (actions) `_init-and-lint.yml`, reduce to emoji labels (#846)
Wasting time `init` to use the cache in `lint` here. Keep both for now, but only call `_init-and-lint`

For Vercel, it's as fast as it'll go for now.

The way to improve this further would be to not concern ourselves with getting the success back from Vercel, and instead either skip that wait, or find a way to eject the wait.

The value in us doing this here again is that we do not want to do `lint/ test` on Vercel, but we also do not wan to create a deployment if there is an obvious problem.