👷  introduce github actions over `_workflows` :octocat:  (#847)
This is not _that much_ faster, which begs the question why. However, it was good to learn how to do this I feel, heh.

Between this and #846 there are a few ways to do this now.

Would really like to eject away from waiting for Vercel's success message. It would probably be more ideal to introduce when that build starts from GitHub rather than calling it from GitHub. IF that makes any sense.