⚡️  header, footer, img quality, `fast-json-stable-stringify` (#858)
- ⚡️  (header) fresnel for menu (footer) do not ssr
- ➕  `fast-json-stable-stringify`
- ⚡️  reduce quality to `80`, add cdn preconnects
- ⚡️ proper code for TextAnnotations, proper NoSSR for index
- ⚡️ Emoji|EmojiParser|Toast eject to custom, dynamic 
- ⬆️ `@jeromefitz/design-system|notion` prerelease update
- ⚡️ uh-oh, haha, looks like Notion content is !SSR now
- 🚑️ SSR EmojiParser for Notion Content, + NoSSR over 10


So this ended up being a big refactor that has been helpful.

But now that some things are distributed, it's not a consistent improvement.

Continue to look into:

- `next/dynamic` vs `NoSSR`:
  - Both are good however for `Emoji|EmojiParser` in particular moving that chunk puts the code elsewhere which makes sense, however, it is adding (I think) to the blocking a tinch (Total Blocking Time: `190ms`).


BEFORE this PR:
```sh
jeromefitzgerald.com:build: ├ ● /[...catchAll] (ISR: 3600 Seconds) (4053 ms)  857 B  212 kB
```

AFTER this PR:
```sh
jeromefitzgerald.com:build: ├ ● /[...catchAll] (ISR: 3600 Seconds) (4208 ms)  64 B  178 kB
```
 
One would think this would be a slam dunk in making the page load faster. 😬 

We trimmed `34 kB` off of the `First Load JS`, but we may have just moved it to a chunk that is no longer in the first load, but needs to be. 🤔 

Truly this is a massive file all things considered and is doing very well regardless:


More stats (before moving to Production Servers):

- Performance: `96`, `86`, `87`
- FCP: `0.3s`, `0.32`, `0.4s`
- Speed Index: `0.7.s`, `3.9s`, `0.9.s`
- LCP: `0.5s`, `0.5s`, `0.4s`
- TTI: `1.4s`, `1.2s`, `1.6s`
- TBT: `190ms`, `100ms`, `360ms`
- CLS: `0`, `0`, `0`

So really the thing we need to look into now is why does the `Speed Index` and `TBT` vary so much?