💚  (actions) 🍽️  of `deploy-to-vercel-action` test (#887)
Previously if the `commit.message` of the `SHA` was very long:

- `pull_request`: with a lot of information the commit message (dependabot, feature, etc.)

This was creating a very large `githubCommitMessage` environment variable which was breaking (specifically) production builds from `main`:

```sh
Error! The total size of all Environment Variables (9.1KB) exceeds 4KB.
[51](https://github.com/JeromeFitz/websites/runs/5115282111?check_suite_focus=true#step:6:51)
Learn More: https://vercel.link/env-vars-limit
```

Being that there are less than 10 environment variables for this website, the hope is this was the culprit.

I am also adding a lot of information in this PR to ensure we are well over the `50` character limit which is now imposed by it truncating in the forked repo.

If this works (when? wishful thinking) then we'll create a PR to see if it may be helpful at all.