♻️  more prep for ISR revalidation away from double cache
The revalidation is not working currently so not adding it.
This stops the double cache of JSON + REDIS

Prioritize REDIS on BUILD, and then post BUILD prioritize
 direct API instead of CACHE

Will most likely introduce `swr` more for particular routes.
 Though ISR would be preferable.

That way when the CMS updates, this could tell Next.
Instead of relying on SWR

Wh00 boy this is a cluster ain't it haha