♻️  `getStaticPropsCatchAll` chaos before peace apparently
Taking a break, haha.

When this is all said and done it will be cool.
Until then, utter chaos as I go through the whiteboard diagrams haha

`getStaticPropsCatchAll` will remove and replace:
- getDataReturn
- getCatchAll

And then we will create its own sub-funcdtions

And then LIFT cache away, though to be honest, maybe that is much later. As for
all intents and purposes I am the only developer using this at the moment.

The goal is to run everytihng buts its cache key so you can pick your choice:
- swr
- zustand
- redis
- etc.

And just go from there. With ISR this seems all moot now, however, the real goal
here is to continue to not rely on Notion's API for the 15-20 builds a day as
thye are not honoring 200 requests or so, and then on the secret project they
are really not vibing with the rest.

This continues to be the proof of concept project,a nd for that we are thankful.