♻️  (next-notion) prep for notionhq/client@2 [skip ci] (#1209)
### Overview

Would be ideal if instead of doing any html formatting within `next-notion` if it churned out the adjusted datapoints / nodes. Right now a lot of the `./src/app/components/...` is nice but locks in even basic design choices.

The packages need to be adjusted before publishing as it is needlessly large at the moment. 

Hard-coded values need to get updated away (specifically for SEO and Podcast stuff)

### useID

`uuid` should be removed and swapped for `React.useID()`

**Update:** Hrm, `getContentNodes` may not be able to do that as this is not a FC per se.

### peerDependencies

- [x] `ioredis` (especially if cacheType will not be redis)
- [x] `plaiceholder`

### Commits

- 📦 dependencies refactor (add/remove/peer)