🚑️  (seo) NICE-18 root `not-found` is breaking seo [b] (#1530)
This is pretty wild. But no time to really look into this. Noticed that the SEO/OG was not rendering but Lighthouse and Analytics seemed to be okay. But this is obviously _not good_. 😆 

Chucking `not-found` into a `@temp` directory.

Probably need to a do a check on any routes that did not get moved over in the `v13` too. But that is not as big of a priority as this is more of an immediate fix thing.

Reference: `https://github.com/vercel/next.js/issues/46619#issuecomment-1475379813`