🩹  (next) NICE-17 revalidate prep app + isr... [b]
This does not seem to work really
 well. From the issues/discussions seems like this is not
 ready for `./app`. So this is kind of half hack, half-ISR
 which will be replaced at some point with: `revalidatePath()`

Getting word that this works on Vercel Deployments,
 not locally. This is kind of true.

Locally, I do not have it working with: `next@13.2.5-canary.17`
- I get an error, but it still does not work when you go to the what
  should be updated page
- Does not show in a refresh either

But if I downgrade to `next@13.2.3`, I get an error -- but it works.
 - Works meaning without a refresh the data is updated on the website

Curious. So going to see if this does "work" on a Vercel Deployment.

if so -- great! If not -- may have to downgrade