🩹  (next) NICE-17 lazy (!dynamic); priority image from cms (#1534)
## Image

Okay the actual fix for the image/template/LCP+CLS was moving away from:

```tsx
import dynamic from 'next/dynamic'
```

And to:

```tsx
import { lazy } from 'react'
```

Not an easy thing to figure out, but that is the point of beta and canary 😅 
Very easy fix once could go through debugging.

## ISR

This is "working" and "not working" now. We do get a `200 Success` on ISR Revalidation, the Redis DB is updated, but `next` is still holding tight to the original value at build time.

Makes sense as I think we are wrapping a lot in `cache()` throughout.

Not a priority but have kept updates as-is as we continue to work towards getting that to work on-demand.

## Upgrades

Package upgrades as there was a build error the past few weeks in `packages` apparently. Oh well, heh.

Needed to get in a fix to `@jeromefitz/notion` to not required `hasChild` (`?.hasChild`).