✨ NICE-23 app dir to use api not direct functions (#1554)
This _was_ working a-okay, but in a `canary` build this way of getting data within the app dir went the way of the `dodo`.

- https://nextjs.org/docs/messages/deopted-into-client-rendering

The reference hyper-focuses on `useSearchParams` but there is a _bit_ more to it. At least in the way this repo had things set up. I think it could also be the way it was doing a lot more than it should have and duplicating the `api` calls. However, originally, I thought that was the preference for RSC in that you did _not_ have to do an API call from RSC.

I digress. This is working much better, so this refactor (aka feature) is a stepped improvement for sure. At one point an update back on `canary` everything "looked good" but **tanked** RES (Real Experience Score). The Lighthouse was still `100` so this was weird.

Added bonus `ISR` now works again by following the `fetch` pattern more closely. And a proper `preload`.

- [x] upgrade to `next@13.3.0`
- [x] upgrade to `@vercel/og@0.5.2`, `swr@2.1.3`
- [x] Try and find what is causing the deopt
  - This was **three fold** at least: 1) `Analytics` 2) `CommandMenu` 3) `getNotionData`
- [x] Trial and Error on `colophon` so that it does not show: `page deopted into client-side rendering`
- [x] Migrate all `page.tsx` to this and update the Page/ Metadata/ Preload
- [x] Verify ISR
- [x] Determine if this works on Vercel

Yes the big callout is a bit of a hack. Since we _use_ the Next API to generate itself, this is probably anti-pattern. (Lol, this whole websites repo has been from day one.)

### `sites/jeromefitzgerald.com/src/lib/constants.ts`

This is a bit of a headache to look at so at some point we should refactor. This _requires_ the Production API to already exist. So this would work for this website, but not a net-new one. So need to also add to README or the call out in a comment here. I think this is kind of my problem with this site using Notion. Perhaps ... and hear me out here ... it was never a good idea to do this and spend this much time on it haha.

- Notion API is _a lot_
- We take the data and customize ("normalize") it to be digestible within this app an store the output in Redis
- We then use the Next API to call Redis and add some realtime flair from time to time

All of that is overcomplicated and probably not necessary. Or can be improved so let us first **fix** this and get back on canary latest and then see when (if ever) there is time to get out of this quagmire.