🚧 NICE-39 design-system 2.0.0-canary.1 (internal for now) 🐦 (#1606)
Need to break this ticket up into many, many small ones as this is a bit of a learning curve and giant undertaking.

Since `tailwind` removes any unused stuff, want to rework a bit to pass _all_ `radix` type colors to it. Down the line you can limit as needed but for now this is just a test of sorts.

Want to move components out to a localized `@jeromefitz/ds` short-term as we look to continue to streamline everything so we can server multiple websites.

Need to merge this WIP as it is holding up a higher priority item which is a Notion rewrite, lol. But I finally figured out why Formulas are good. 🤣