🧨 (revamp) NICE-43 Next RSC, Notion API, & Tailwind [b] [!] (#1621)
# ⚠️ Please Note

This is a bit bigger than I would want, but this is a complete refactor to better utilize a `Notion CMS` within a `Next 13 App` structure taking advantage of `RSC`. And then for good measure started the migration _back_ to `Tailwind` (but now I think `Stitches|Vanilla Extract` are oaky with Next 13).

✨ NICE-43 revamp Next RSC, Notion API, & Tailwind

BREAKING CHANGE: New Notion Implementation through Next 13 RSC (App)

# Complete Refactor for Next RSC

- [x] Redis KV Cache
  - Good for quicker Vercel builds / and reducing Notion random limits
  - Next will still limit due to deduplicated requests
    - Sidenote: Need to get better at _not_ prop-drilling and just calling function again
- [x] Image Component
  - [x] Next Refactor
  - [x] Notion Refactor
- [x] Tailwind
  - [x] Move away from Stitches
  - [x] Temporary move to `@jeromefitz/ds` localized version
  - [x] Radix UI Refactor
  - [x] Use internal Component Library for now
- [x] Notion Refactor

  - [x] Image: Hosted by Notion Expiration Refresh
  - [x] Rely less on custom queries and manipulation, more on out of the box Notion API, React, and Next
    - [x] Move from `nodes` to more straight-forward components
    - [x] Refactor logic for
      - [x] `Column|ColumnList`
      - [x] `ListBulleted|ListItem|ListNumbered`
      - [x] `TextAnnotations`
    - [x] **New**: `Embed|Video`
      - [x] `Embed.Twitter` <= Twitter is self-imploding, so good timing.
      - [x] `Embed.YouTube` <= YouTube is ... no comment.

- [x] Next
  - [ ] `preload` => hold on this for now not working as expected
  - [x] `generateMetadata` => move away from `next-seo` (rip)
    - [ ] Need to re-incorporate `@vercel/og` -- or not temporarily.
  - [x] `generateStaticParams` => hack if `isDev` to stop running all the time :?
  - [x] `not-found` => Instead of customizing every route, use `notFound()`
  - [x] `robots.ts|sitemap.ts` => move away from `next-sitemap` (rip)
- [x] Package Upgrades
  - [x] Basically everything except `plaiceholder|semantic-release|syncpack`
- [x] Other refactors
  - [x] EmojiWrapper
  - [x] Config Packages
    - [x] `./packages/*` – this all needs to be ported to [`@jeromefitz/packages`](https://github.com/JeromeFitz/packages)
      - Was borderline impossible to do these in-tandem (need to get better at `pnpm linking`)
- [x] Fathom advises against custom domains now :/

## Layout

Not to be lost in the backend type stuff but this is a complete rehaul of presentation.

The eventual goal of this `Notion + Next` implementation is you can take the data in any way and display it as you would like. For now though still tightly coupled with a lot of decisions until I can abstract away further. Which -- at teh rate I am going may never happen.

## Notion

Do not "normalize" data from Notion, embrace it.

- This is the CMS why are we going through all the trouble to hack it
- Create new Block Components to match and mimic needs

### Formulas

The "big change" here is utilizing Notion Formulas in the CMS.

#### Slug

Before we had `Slug` now we populated `Slug.Preview` where we are able to have Notion mimic Routes better Server Side than figuring it all out on the fly in `next`.

**Note:** For `Events` we need to manually write the `Slug` for now. The potential clash of having multiple Headline Acts means we could have a url like:

- `../jerome-and,alex-o-jerome`

Which I think _could_ work because we could loop through but also _very_ confusing as a URL to share, haha.

To recap:

- `Slug.Preview` for example will create the matching Next Segment Route
  - Before: `jerome-and` (shows); After: `/shows/jerome-and`
  - Before: `jerome-and` (events); After: `/events/2023/06/16/jerome-and`
  - Before: `homepage` (pages); After: `/`

##### Date

Same here, instead of getting the data from Notion then running through `date-fns`, Notion uses `moment` (I think). So we can do more preparation which makes for a more straight-forward:

```bash
- Date: July 15, 2023 9:00PM
- Date.DayOfMonth: 15
- Date.DayOfMonthOrdinal: 15th
- Date.DayOfWeek: Saturday
- Date.DayOfWeekAbbr: Sat
- Date.DaysUntilEvent: 15
- Date.HoursUntilEvent: 360
- Date.ISO: 2023-07-15T21:00:00-04:00
- Date.Month: 07
- Date.MonthName: July
- Date.MonthNameAbbr: Jul
- Date.Time: 09:00 PM
- Date.Timezone: EDT
- Date.WeekNumber: 28
- Date.Year: 2023
```

## Next

### App

The gains of doing `[[...catchAll]]` for everything at **root** was not worth it. 😅

Especially if we _want_ to be able to change layout, RSC data points and more based on the Segment (Notion Database Type).

So there is some duplication of code throughout but will look into better ways of lifting the "same" stuff.

### Events

Ability to see a range of events depending on Date (say for a week, or a weekend):

- `events/yyyy/mm/dd/to/yyyy/mm/dd`
- `events/2023/06/29/to/2023/07/02`

Props to Katie T. as we came up with this idea in a brainstorm session.

- `events/[from]/to/[to]`
- `events/2023/06/15/to/2023/06/18` (THU-SUN)
- `events/2023/06/29/to/2023/05/01` (Across Months)
- `events/2023/12/28/to/2025/01/07` (Across Years)

Probably need to do something for malformed dates, but I reckon `404` will do that.

Up next would be ideas for `Landing Pages`.

### In-House

`next-seo` and `next-sitemap` will eventually be absorded, or their functionality at least, within `next` itself. This gets the ball rolling on that. They were awesome, and thank you.

Depending on your use case those will still be very valid for use!

## Packages

This just covers the `sites/jeromefitzgerald.com` for now

**Upgrade:**
```bash
@mantine/hooks@6.0.15
@notionhq/client@2.2.6
@radix-ui/...@latest
@upstash/redis@1.22.0
@vercel/og@0.5.8
framer-motion@10.12.18
next@13.4.8-canary.14
react-swipeable@7.0.1
swr@2.2.0

@types/ramda@0.29.3
@types/react@18.2.14
@types/react-dom@18.2.6
@types/uuid@9.0.2
tailwind-merge@1.13.2
```

**Add:**

```bash
emoji-regex@10.2.1
node-emoji@2.1.0
react-headroom@3.2.1 # use framer-motion instead
react-tweet@2.0.2 # 3.x has no release notes

@types/react-headroom@3.2.0
```

**Remove:**

```bash
next-seo
next-sitemap
next-unused
```