🩹  (deps) NICE-79 radix-ui/themes force esm [b] (#2266)
- [x] 🩹  (deps) NICE-79 radix-ui/themes force esm

This is a temporary patch to see about _ensuring_ `@radix-ui/themes` uses **esm**.

This does enable tree-shaking to an extent on the build. 

(Still work to do with the `.css` import to limit un-used components. But is not a priority.)

## Edit

This may be solely because of the `tsconfig` locally too. 🤔

- I believe so because was not able to move `next` awhile back to `NodeNext`
- This will be a hack until I can revisit that probably

---
Additional Update while we are here

- [x] 🌐 (seo) fix `music` 🚚 `Top` to `Music.client`