🩹  (radix-ui) NICE-88 37.1 KiB CSS Reduction test [b]
This is from the branch:

- `sandbox/color-reorg`

Trying out ability to only choose which CSS is generated for:

- colors
- components

Most applications may not use all of them.

Currently the build for themes is one CSS file

- Probably better to figure out alternatives to trimming
- Rather than this POC (POS?) attempt 🤣

Stats:

- Before: `75.1 KiB`
- After: `38.0 KiB`

This is with a few extra colors that may be able to be removed.

Though the tokens are not really the big deal here (to me).

If we could pass the `.css` file through next/cssnano/brotli 🤔

- Could/would negate all this ➕