🩹  NICE-93 images for books ♻️ storybook [b] (#2327)
Update API to pull Images and Bookshop URL from Notion.

- This is still hacky as I did not want to spend the time to apply teh AWS Expiration from Notion here, and instead hosted images on a CDN and do not feel like redoing all these Book entries 😂


Also updates `storybook` a bit to be `ts` --- but most likely Storybook is going away, I do not use it and have moved away from a custom Design System to use `@radix-ui/themes` ✅ 

Introduces a dropdown to limit what is shown to someone on the init.

- Eventually this should pull dynamically from an API like Music
- No need to do this now of course

**Commits:**

- [x] 🩹  NICE-93 images for books
- [x] 🚚 (storybook) NICE-93 jsx(x)=>ts(x)
- [x] ♻️  NICE-93 add bookStatus to store