🩹  (radix-ui) NICE-97 reduced `Anchor` 🔗 [b] (#2398)
Move `AnchorUnstyled` _to be_ `Anchor`.

----

No idea why this does not happen on localized build but does on hosting.
----

Current

```tsx
import { Flex } from '@radix-ui/themes'
```

- PRO: More legible; Less up-keep on `dist` changes
- CON: !tree-shaking

Preferred

```tsx
import { Flex } from '@radix-ui/themes/dist/esm/components/flex.js'
```
- PRO: Tree-shaking (`49 KiB` savings in unused JavaScript)
- CON: Have to call every component like this to take advantage

---

If we call this directly (`./dist/esm/components/flex.js`) there is an intermittent error triggered via routing.

This looks to be due to `flex.props` and the `parseJustifyValue` being created _after_ it is called.
- Which I do not htink _should_ be a problem, but perhaps with the way this is minified / prepared for production does matter for the order of the operations

This PR is more information gathering to be able to explain this issue and see if there is a fix,
 or instead shelve and point to this as a discussion point for a later date.

 ----

 **Update:** I actually think this is a routing issue, not a `radix-ui/themes` one. It just looks like it because it is triggering `404`. Not sure. If anything a new problem has come up 😆
♻️ (react) unnecessary calls