👷  pnpm/action-setup@v3 ⬆️ pnpm@9.0.4  🔧 (renovate) !pnpm (#2477)
After reading a lot of issues and pull requests, going to respectfully back away from `corepack` a bit with `github-actions`. (Maybe altogether. 😬 )

- [x] Swap _back_ to `pnpm/action-setup` as `v3` now supports `node@20` (`2024-02-08`)
- [x] Remove `pnpm` from `renovate` (for now?)

Ideally we would keep `pnpm` in-sync across projects. But the root `package.json` can drive `CI` and then locally as long as people are using the same _major_. I don't know, so many opinions. 😆 

## Previous Commits:

A few months ago when `pnpm/action-setup@v2.4.0` was not available with `node@20` there was a custom install.

This PR did originally build on that to pull from `packageManager` via `./package.json`:

```yaml
    - name: '🔧  pnpm: Get Corepack Version'
      id: node-pnpm-version
      shell: bash
      run: echo PNPM_VERSION=$(node --print 'require("./package.json").packageManager') >> $GITHUB_ENV

    - name: '🔧  pnpm: Install'
      id: pnpm-setup
      shell: bash
      # @note(corepack) Why, that was quite the rabbit hole
      #  on corepack|npm and github actions ... yikes.
      #
      # How about you just install it and do not use corepack for now.
      # :homer-simpson-fade-backwards-into-bushes:
      #
      run: |
        npm install -g ${{ env.PNPM_VERSION }}
```

Kept here in case we need to do this again for `node@22` if we are ahead.