🚨 NICE-98 eslint@9|jeromefitz/eslint-config@3 🐦 [skip ci] (#2479)
**Disclaimer:**

Eh, should wait until this is off of canary. But want to see if this works not on my machine.

**Overview:**

A way too early move to `eslint@9` considering it will take a few months for the ecosystem to upgrade to flat configuration and more. This does some hacks (surprises) to learn a bit more of the new format and probably will get some tweaks here and there to some sort of actual standard at some point.

Things that are not working:

- react/exhaustive-deps
- `@next/next/...` certain ones
- `testing-library`
- `storybook`

All good.